♻️ app-writer.GrapesEditorState => separate code related to grapes' plugins
