+++ src/app/app-writer/grapes-editor/index.ts
@@ -2,6 +2,7 @@ export * from './grapes.config'
 export * from './grapes.state'
 export * from './grapes.storage'
 export * from './grapes.view'
+export * from './grapes.plugins'
 export * from './manager-style'
 export * from './toolboxes.view'
 export * from './utils.view'
